qemu: split config and config tests into their own files to match best practices from other builders
